fix some bugs
修复了数值计算中可能存在的溢出和均值算法问题

+++ EmbeddedProject/APP/src/monitor.h
@@ -9,6 +9,6 @@
 
 void Monitor_PWM_Init();
 
-void Monitor_Set_Usage(uint8_t ch, uint8_t usage);
+void Monitor_Set_Usage(uint8_t ch, uint16_t usage);
 
 #endif //CH571PRJ_MONITOR_H
